chore: restructuring of routing
