+++ app/(tabs)/_layout.tsx
@@ -1,43 +1,19 @@
-import FontAwesome from "@expo/vector-icons/FontAwesome";
-import { Redirect, Tabs } from "expo-router";
-import { getAsyncUser, useAuthStore } from "../../src/utils/store";
-import { useEffect } from "react";
-
-/**
- * You can explore the built-in icon families and icons on the web at https://icons.expo.fyi/
- */
-function TabBarIcon(props: {
-  name: React.ComponentProps<typeof FontAwesome>["name"];
-  color: string;
-}) {
-  return <FontAwesome size={28} style={{ marginBottom: -3 }} {...props} />;
-}
+import { Redirect } from "expo-router";
+import { Drawer } from "expo-router/drawer";
+import { Platform } from "react-native";
+import { LeftSidePanel } from "../../src/features/layouts/LeftSidePanel";
+import { Tabs } from "../../src/features/layouts/Tabs";
+import { useAuthStore } from "../../src/utils/store";
 
 const TabLayout = () => {
   const user = useAuthStore((state) => state.user);
   if (!user) {
     return <Redirect href="/(auth)/login" />;
   }
-  return (
-    <Tabs>
-      <Tabs.Screen
-        name="index"
-        options={{
-          title: "Home",
-          tabBarIcon: ({ color }) => <TabBarIcon name="home" color={color} />,
-          headerShown: false,
-        }}
-      />
-      <Tabs.Screen
-        name="profile"
-        options={{
-          title: "Profile",
-          tabBarIcon: ({ color }) => <TabBarIcon name="user" color={color} />,
-          headerShown: false,
-        }}
-      />
-    </Tabs>
-  );
+
+  if (Platform.OS === "web") return <Drawer drawerContent={LeftSidePanel} />;
+
+  return <Tabs />;
 };
 
 export default TabLayout;
