+++ app/(tabs)/home/bookSearch.tsx
@@ -1,0 +1,5 @@
+import { BookSearchScreen } from "../../../src/features/books/screens/BookSearchScreen";
+
+export default function BookSearch() {
+  return <BookSearchScreen />;
+}
