+++ app/(tabs)/home/index.tsx
@@ -1,0 +1,5 @@
+import { HomeScreen } from "../../../src/features/home/HomeScreen";
+
+export default function Home() {
+  return <HomeScreen />;
+}
