--- app/(tabs)/index.tsx
@@ -1,8 +1,0 @@
-import { HomeScreen } from "../../src/features/home/HomeScreen";
-
-export default function Home() {
-  return (
-    <HomeScreen />
-  );
-}
-
