--- app/(tabs)/profile.tsx
@@ -1,5 +1,0 @@
-import { ProfileScreen } from "../../src/features/profile/screens/ProfileScreen";
-
-export default function Profile() {
-  return <ProfileScreen />;
-}
