+++ app/(tabs)/profile/index.tsx
@@ -1,0 +1,5 @@
+import { ProfileScreen } from "../../../src/features/profile/screens/ProfileScreen";
+
+export default function Profile() {
+  return <ProfileScreen />;
+}
