--- app/bookSearch.tsx
@@ -1,5 +1,0 @@
-import { BookSearchScreen } from "../src/features/books/screens/BookSearchScreen";
-
-export default function BookSearch() {
-  return <BookSearchScreen />;
-}
