+++ package.json
@@ -13,27 +13,29 @@
     "preset": "jest-expo"
   },
   "dependencies": {
-    "@expo/vector-icons": "^14.0.0",
+    "@expo/vector-icons": "^14.0.2",
     "@firebase/auth": "^1.3.0",
     "@react-native-async-storage/async-storage": "1.23.1",
+    "@react-navigation/drawer": "^6.6.15",
     "@react-navigation/native": "^6.0.2",
     "axios": "^1.5.0",
     "babel-preset-expo": "^11.0.5",
-    "expo": "~51.0.2",
-    "expo-font": "~12.0.4",
+    "expo": "~51.0.21",
+    "expo-font": "~12.0.9",
     "expo-linking": "~6.3.1",
-    "expo-router": "~3.5.11",
-    "expo-splash-screen": "~0.27.4",
+    "expo-router": "~3.5.18",
+    "expo-splash-screen": "~0.27.5",
     "expo-status-bar": "~1.12.1",
-    "expo-system-ui": "~3.0.4",
+    "expo-system-ui": "~3.0.7",
     "expo-web-browser": "~13.0.3",
     "firebase": "^10.3.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-native": "0.74.1",
+    "react-native": "0.74.3",
     "react-native-dotenv": "^3.4.9",
     "react-native-gesture-handler": "~2.16.1",
-    "react-native-safe-area-context": "4.10.1",
+    "react-native-reanimated": "~3.10.1",
+    "react-native-safe-area-context": "4.10.5",
     "react-native-screens": "3.31.1",
     "react-native-web": "~0.19.10",
     "zustand": "^4.4.1"
@@ -42,7 +44,7 @@
     "@babel/core": "^7.24.0",
     "@types/react": "~18.2.79",
     "jest": "^29.2.1",
-    "jest-expo": "~51.0.1",
+    "jest-expo": "~51.0.3",
     "react-test-renderer": "18.2.0",
     "typescript": "~5.3.3"
   },
