+++ src/features/layouts/Tabs.tsx
@@ -1,0 +1,33 @@
+import FontAwesome from "@expo/vector-icons/FontAwesome";
+import { Tabs as ExpoTabs } from "expo-router";
+import React from "react";
+
+function TabBarIcon(props: {
+  name: React.ComponentProps<typeof FontAwesome>["name"];
+  color: string;
+}) {
+  return <FontAwesome size={28} style={{ marginBottom: -3 }} {...props} />;
+}
+
+export const Tabs = () => {
+  return (
+    <ExpoTabs>
+      <ExpoTabs.Screen
+        name="index"
+        options={{
+          title: "Home",
+          tabBarIcon: ({ color }) => <TabBarIcon name="home" color={color} />,
+          headerShown: false,
+        }}
+      />
+      <ExpoTabs.Screen
+        name="profile"
+        options={{
+          title: "Profile",
+          tabBarIcon: ({ color }) => <TabBarIcon name="user" color={color} />,
+          headerShown: false,
+        }}
+      />
+    </ExpoTabs>
+  );
+};
